去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@22778 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/rtn/rtnCoordCommon.cpp
@@ -1673,7 +1673,7 @@ namespace engine
                PD_LOG ( PDWARNING, "Received invalid msg-event(data is null)" );
                continue ;
             }
-            if ( (UINT32)pMsg->opCode != MSG_CAT_GRP_RES ||
+            if ( pMsg->opCode != MSG_CAT_GRP_RES ||
                  pMsg->requestID != reqID )
             {
                if ( cb->getCoordSession()->isValidResponse( reqID ) )
